Add infrastructure for smarter sentence models
The idea behind these additions is that each sentence is modeled by a
Sentence type, which will maintain information about the keyword of the
sentence, as well as data related to that keyword. The data is
maintained by the new DataType class, which is capable of imposing
restrictions on the data.

For example, there is a TimeData DataType, which raises an exception if
any data not of the form "11:00 AM" is passed to it. Additionally,
DataTypes may be constructed with a maximum number of data elements.