Do not normalize Sentence.primary
Sentence.primary is used for as a human-readable or preferred name for
the Keyword. For example, we might refer to a "start time" Keyword that
has many alternatives such as "meeting time" or "starts". "start time"
would be the primary expression, and should not be normalized, but the
list of words used to compare (which includes a copy of the primary
word) is normalized.